move ManifestViewUtils and ResourceViewUtils to viewmodel package
