+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/activity/TractActivity.java
@@ -54,10 +54,10 @@ import org.cru.godtools.tract.analytics.model.TractPageAnalyticsScreenEvent;
 import org.cru.godtools.tract.content.TractManifestLoader;
 import org.cru.godtools.tract.model.Card;
 import org.cru.godtools.tract.model.Manifest;
-import org.cru.godtools.tract.model.ManifestViewUtils;
 import org.cru.godtools.tract.model.Page;
 import org.cru.godtools.tract.util.DrawableUtils;
 import org.cru.godtools.tract.util.ViewUtils;
+import org.cru.godtools.tract.viewmodel.ManifestViewUtils;
 import org.cru.godtools.tract.widget.ScaledPicassoImageView;
 import org.greenrobot.eventbus.EventBus;
 import org.greenrobot.eventbus.Subscribe;
