+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/model/Card.java
@@ -19,6 +19,7 @@ import org.cru.godtools.tract.R;
 import org.cru.godtools.tract.R2;
 import org.cru.godtools.tract.model.AnalyticsEvent.Trigger;
 import org.cru.godtools.tract.model.Page.PageViewHolder;
+import org.cru.godtools.tract.viewmodel.ResourceViewUtils;
 import org.cru.godtools.tract.widget.TractPicassoImageView;
 import org.xmlpull.v1.XmlPullParser;
 import org.xmlpull.v1.XmlPullParserException;
