+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/model/Image.java
@@ -10,6 +10,7 @@ import org.ccci.gto.android.common.picasso.view.PicassoImageView;
 import org.ccci.gto.android.common.util.XmlPullParserUtils;
 import org.cru.godtools.tract.R;
 import org.cru.godtools.tract.R2;
+import org.cru.godtools.tract.viewmodel.ResourceViewUtils;
 import org.xmlpull.v1.XmlPullParser;
 import org.xmlpull.v1.XmlPullParserException;
