+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/viewmodel/ManifestViewUtils.java
@@ -1,8 +1,9 @@
-package org.cru.godtools.tract.model;
+package org.cru.godtools.tract.viewmodel;
 
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
+import org.cru.godtools.tract.model.Manifest;
 import org.cru.godtools.tract.widget.ScaledPicassoImageView;
 
 public class ManifestViewUtils {
