+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/viewmodel/ResourceViewUtils.java
@@ -1,4 +1,4 @@
-package org.cru.godtools.tract.model;
+package org.cru.godtools.tract.viewmodel;
 
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
@@ -9,6 +9,9 @@ import android.widget.ImageView;
 import android.widget.RelativeLayout;
 
 import org.ccci.gto.android.common.picasso.view.PicassoImageView;
+import org.cru.godtools.tract.model.ImageGravity;
+import org.cru.godtools.tract.model.ImageScaleType;
+import org.cru.godtools.tract.model.Resource;
 import org.cru.godtools.tract.widget.ScaledPicassoImageView;
 
 import jp.wasabeef.picasso.transformations.CropTransformation;
